debugging ipc with tsx not working

+++ src/child.ts
@@ -10,7 +10,10 @@ export class Child {
   public lastPing: number;
   public softTimeout: Duration;
   public hardTimeout: Duration;
-  public softTimeoutReached: boolean = false;
+  public softTimeoutReached = false;
+
+  private _sigtermSent = false;
+  private _sigquitSent = false;
 
   constructor(wakaq: WakaQ, process: ChildProcess) {
     this.process = process;
@@ -20,7 +23,9 @@ export class Child {
   }
 
   public sigterm() {
+    if (this._sigtermSent) return;
     this.process.kill('SIGTERM');
+    this._sigtermSent = true;
   }
 
   public sigkill() {
@@ -28,7 +33,9 @@ export class Child {
   }
 
   public sigquit() {
+    if (this._sigquitSent) return;
     this.process.kill('SIGQUIT');
+    this._sigquitSent = true;
   }
 
   public setTimeouts(wakaq: WakaQ, task?: Task, queue?: WakaQueue) {
